# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Yoko API Documentation",
        "version": "1.0"
    },
    "paths": {
...
        "/api/v1/auth/register": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Register users",
                "operationId": "5aff8982c53bc6d49f34b9064d057b85",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "allOf": [
                                    {
                                        "properties": {
                                            "name": {
                                                "type": "string",
                                                "example": "Maksym"
                                            },
                                            "surname": {
                                                "type": "string",
                                                "example": "Prokopenko"
                                            },
                                            "email": {
                                                "type": "string",
                                                "example": "MaksymProkopenko@gmail.com"
                                            },
                                            "password": {
                                                "type": "string",
                                                "example": "password"
                                            }
                                        },
                                        "type": "object"
                                    }
                                ]
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Request successful",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "Request successful"
                                        },
                                        "message": {
                                            "type": "string",
                                            "nullable": true
                                        },
                                        "data": {
                                            "properties": {
                                                "user": {
                                                    "properties": {
                                                        "id": {
                                                            "type": "integer",
                                                            "example": 1
                                                        },
                                                        "email": {
                                                            "type": "string",
                                                            "example": "MaksymProkopenko@gmail.com"
                                                        },
                                                        "name": {
                                                            "type": "string",
                                                            "example": "Maksym"
                                                        },
                                                        "surname": {
                                                            "type": "string",
                                                            "example": "Prokopenko"
                                                        }
                                                    },
                                                    "type": "object"
                                                },
                                                "token": {
                                                    "type": "string",
                                                    "example": "4|c6u23dX8dk3A4lxDQpdfO4oNclfbneljphpw5xZN9c1728e7"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/auth/login": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Login for users",
                "operationId": "2f2f70d659c197d0a78e19c22365827a",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "allOf": [
                                    {
                                        "properties": {
                                            "email": {
                                                "type": "string",
                                                "example": "MaksymProkopenko@gmail.com"
                                            },
                                            "password": {
                                                "type": "string",
                                                "example": "password"
                                            }
                                        },
                                        "type": "object"
                                    }
                                ]
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Request successful",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "Request successful"
                                        },
                                        "message": {
                                            "type": "string",
                                            "nullable": true
                                        },
                                        "data": {
                                            "properties": {
                                                "user": {
                                                    "properties": {
                                                        "id": {
                                                            "type": "integer",
                                                            "example": 1
                                                        },
                                                        "email": {
                                                            "type": "string",
                                                            "example": "MaksymProkopenko@gmail.com"
                                                        },
                                                        "name": {
                                                            "type": "string",
                                                            "example": "Maksym"
                                                        },
                                                        "surname": {
                                                            "type": "string",
                                                            "example": "Prokopenko"
                                                        }
                                                    },
                                                    "type": "object"
                                                },
                                                "token": {
                                                    "type": "string",
                                                    "example": "4|c6u23dX8dk3A4lxDQpdfO4oNclfbneljphpw5xZN9c1728e7"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/auth/admin-login": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Login for admins",
                "operationId": "cc9f35eb743ca6c2b432b2d7bc542b0e",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "allOf": [
                                    {
                                        "properties": {
                                            "email": {
                                                "type": "string",
                                                "example": "admin@admin.com"
                                            },
                                            "password": {
                                                "type": "string",
                                                "example": "password"
                                            }
                                        },
                                        "type": "object"
                                    }
                                ]
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Request successful",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "Request successful"
                                        },
                                        "message": {
                                            "type": "string",
                                            "nullable": true
                                        },
                                        "data": {
                                            "properties": {
                                                "user": {
                                                    "properties": {
                                                        "id": {
                                                            "type": "integer",
                                                            "example": 1
                                                        },
                                                        "email": {
                                                            "type": "string",
                                                            "example": "MaksymProkopenko@gmail.com"
                                                        }
                                                    },
                                                    "type": "object"
                                                },
                                                "token": {
                                                    "type": "string",
                                                    "example": "4|c6u23dX8dk3A4lxDQpdfO4oNclfbneljphpw5xZN9c1728e7"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "apiKey",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header"
            }
        }
    }
}